Backend/requirements.txt
torch==2.1.0
torchvision==0.16.0
pillow==10.0.0
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
pymongo==4.1.1
PyJWT==2.8.0
bcrypt==4.0.1
requests==2.31.0
numpy==1.24.3
ultralytics==8.0.0
timm==0.9.2
transformers==4.35.0
matplotlib==3.8.0